refactor(PositionConverter): change domain name to PositionConverter
